use long instead of int for Stream Position

+++ src/foundation/src/PDFsharp/src/PdfSharp/Pdf.Signatures/PdfArrayWithPadding.cs
@@ -36,11 +36,11 @@ namespace PdfSharp.Pdf.Signatures
         /// <summary>
         /// Position of the first byte of this string in PdfWriter's Stream
         /// </summary>
-        public int PositionStart { get; internal set; }
+        public long PositionStart { get; internal set; }
 
         /// <summary>
         /// Position of the last byte of this string in PdfWriter's Stream
         /// </summary>
-        public int PositionEnd { get; internal set; }
+        public long PositionEnd { get; internal set; }
     }
 }
